chore: reduce circular deps in @grafana/ui
