+++ packages/grafana-ui/src/components/BigValue/BigValue.story.tsx
@@ -4,15 +4,15 @@ import { type FieldSparkline, FieldType } from '@grafana/data';
 
 import { useTheme2 } from '../../themes/ThemeContext';
 
+import { BigValue } from './BigValue';
+import mdx from './BigValue.mdx';
 import {
-  BigValue,
   BigValueColorMode,
   BigValueGraphMode,
   BigValueJustifyMode,
   BigValueTextMode,
   type Props,
-} from './BigValue';
-import mdx from './BigValue.mdx';
+} from './BigValueTypes';
 
 const meta: Meta = {
   title: 'Plugins/BigValue',
